+++ src/backend/core/tests/users/test_api_users_list.py
@@ -8,17 +8,11 @@ import pytest
 from rest_framework.status import (
     HTTP_200_OK,
     HTTP_401_UNAUTHORIZED,
-    HTTP_403_FORBIDDEN,
-    HTTP_405_METHOD_NOT_ALLOWED,
 )
 from rest_framework.test import APIClient
 
 from core import factories, models
-from core.api.client import serializers
 from core.api.client.viewsets import Pagination
-from core.factories import TeamAccessFactory
-
-from mailbox_manager.factories import MailDomainAccessFactory
 
 pytestmark = pytest.mark.django_db
 
@@ -513,427 +507,3 @@ def test_api_users_list_pagination_wrong_page_size(
 
     # Length should not exceed "max_page_size" default value
     assert len(content["results"]) == 100
-
-
-def test_api_users_retrieve_me_anonymous():
-    """Anonymous users should not be allowed to list users."""
-    factories.UserFactory.create_batch(2)
-    client = APIClient()
-    response = client.get("/api/v1.0/users/me/")
-    assert response.status_code == HTTP_401_UNAUTHORIZED
-    assert response.json() == {
-        "detail": "Authentication credentials were not provided."
-    }
-
-
-def test_api_users_retrieve_me_authenticated():
-    """Authenticated users should be able to retrieve their own user via the "/users/me" path."""
-    user = factories.UserFactory(with_organization=True)
-
-    client = APIClient()
-    client.force_login(user)
-
-    # Define profile contact
-    factories.ContactFactory(owner=user, user=user)
-
-    factories.UserFactory.create_batch(2)
-    response = client.get(
-        "/api/v1.0/users/me/",
-    )
-
-    assert response.status_code == HTTP_200_OK
-    assert response.json() == {
-        "id": str(user.id),
-        "name": str(user.name),
-        "email": str(user.email),
-        "language": user.language,
-        "timezone": str(user.timezone),
-        "is_device": False,
-        "is_staff": False,
-        "abilities": {
-            "contacts": {"can_create": True, "can_view": True},
-            "mailboxes": {"can_create": False, "can_view": False},
-            "teams": {"can_create": False, "can_view": False},
-        },
-        "organization": {
-            "id": str(user.organization.pk),
-            "name": user.organization.name,
-        },
-    }
-
-
-def test_api_users_retrieve_me_authenticated_abilities():
-    """
-    Authenticated users should be able to retrieve their own user via the "/users/me" path
-    with the proper abilities.
-    """
-    user = factories.UserFactory()
-
-    client = APIClient()
-    client.force_login(user)
-
-    # Define profile contact
-    factories.ContactFactory(owner=user, user=user)
-
-    factories.UserFactory.create_batch(2)
-
-    # Test the mailboxes abilities
-    mail_domain_access = MailDomainAccessFactory(user=user)
-
-    response = client.get("/api/v1.0/users/me/")
-
-    assert response.status_code == HTTP_200_OK
-    assert response.json()["abilities"] == {
-        "contacts": {"can_create": True, "can_view": True},
-        "mailboxes": {"can_create": True, "can_view": True},
-        "teams": {"can_create": False, "can_view": False},
-    }
-
-    # Test the teams abilities - user is not an admin/owner
-    team_access = TeamAccessFactory(user=user, role=models.RoleChoices.MEMBER)
-    response = client.get("/api/v1.0/users/me/")
-
-    assert response.status_code == HTTP_200_OK
-    assert response.json()["abilities"] == {
-        "contacts": {"can_create": True, "can_view": True},
-        "mailboxes": {"can_create": True, "can_view": True},
-        "teams": {"can_create": False, "can_view": False},
-    }
-
-    # Test the teams abilities - user is an admin/owner
-    team_access.role = models.RoleChoices.ADMIN
-    team_access.save()
-
-    response = client.get("/api/v1.0/users/me/")
-
-    assert response.status_code == HTTP_200_OK
-    assert response.json()["abilities"] == {
-        "contacts": {"can_create": True, "can_view": True},
-        "mailboxes": {"can_create": True, "can_view": True},
-        "teams": {"can_create": True, "can_view": True},
-    }
-
-    # Test the mailboxes abilities - user has no mail domain access anymore
-    mail_domain_access.delete()
-
-    response = client.get("/api/v1.0/users/me/")
-
-    assert response.status_code == HTTP_200_OK
-    assert response.json()["abilities"] == {
-        "contacts": {"can_create": True, "can_view": True},
-        "mailboxes": {"can_create": False, "can_view": False},
-        "teams": {"can_create": True, "can_view": True},
-    }
-
-
-def test_api_users_retrieve_anonymous():
-    """Anonymous users should not be allowed to retrieve a user."""
-    client = APIClient()
-    user = factories.UserFactory()
-    response = client.get(f"/api/v1.0/users/{user.id!s}/")
-
-    assert response.status_code == HTTP_401_UNAUTHORIZED
-    assert response.json() == {
-        "detail": "Authentication credentials were not provided."
-    }
-
-
-def test_api_users_retrieve_authenticated_self():
-    """
-    Authenticated users should be allowed to retrieve their own user.
-    The returned object should not contain the password.
-    """
-    user = factories.UserFactory()
-
-    client = APIClient()
-    client.force_login(user)
-
-    response = client.get(
-        f"/api/v1.0/users/{user.id!s}/",
-    )
-    assert response.status_code == HTTP_405_METHOD_NOT_ALLOWED
-    assert response.json() == {"detail": 'Method "GET" not allowed.'}
-
-
-def test_api_users_retrieve_authenticated_other():
-    """
-    Authenticated users should be able to retrieve another user's detail view with
-    limited information.
-    """
-    user, other_user = factories.UserFactory.create_batch(2)
-
-    client = APIClient()
-    client.force_login(user)
-
-    response = client.get(
-        f"/api/v1.0/users/{other_user.id!s}/",
-    )
-    assert response.status_code == HTTP_405_METHOD_NOT_ALLOWED
-    assert response.json() == {"detail": 'Method "GET" not allowed.'}
-
-
-def test_api_users_create_anonymous():
-    """Anonymous users should not be able to create users via the API."""
-    response = APIClient().post(
-        "/api/v1.0/users/",
-        {
-            "language": "fr-fr",
-            "password": "mypassword",
-        },
-    )
-    assert response.status_code == HTTP_401_UNAUTHORIZED
-    assert "Authentication credentials were not provided." in response.content.decode(
-        "utf-8"
-    )
-    assert models.User.objects.exists() is False
-
-
-def test_api_users_create_authenticated():
-    """Authenticated users should not be able to create users via the API."""
-    user = factories.UserFactory()
-
-    client = APIClient()
-    client.force_login(user)
-
-    response = client.post(
-        "/api/v1.0/users/",
-        {
-            "language": "fr-fr",
-            "password": "mypassword",
-        },
-        format="json",
-    )
-    assert response.status_code == HTTP_405_METHOD_NOT_ALLOWED
-    assert response.json() == {"detail": 'Method "POST" not allowed.'}
-    assert models.User.objects.exclude(id=user.id).exists() is False
-
-
-def test_api_users_update_anonymous():
-    """Anonymous users should not be able to update users via the API."""
-    user = factories.UserFactory()
-
-    old_user_values = dict(serializers.UserSerializer(instance=user).data)
-    new_user_values = serializers.UserSerializer(instance=factories.UserFactory()).data
-
-    response = APIClient().put(
-        f"/api/v1.0/users/{user.id!s}/",
-        new_user_values,
-        format="json",
-    )
-
-    assert response.status_code == HTTP_401_UNAUTHORIZED
-    assert response.json() == {
-        "detail": "Authentication credentials were not provided."
-    }
-
-    user.refresh_from_db()
-    user_values = dict(serializers.UserSerializer(instance=user).data)
-    for key, value in user_values.items():
-        assert value == old_user_values[key]
-
-
-def test_api_users_update_authenticated_self():
-    """
-    Authenticated users should be able to update their own user but only "language"
-    and "timezone" fields.
-    """
-    user = factories.UserFactory()
-
-    client = APIClient()
-    client.force_login(user)
-
-    old_user_values = dict(serializers.UserSerializer(instance=user).data)
-    new_user_values = dict(
-        serializers.UserSerializer(instance=factories.UserFactory()).data
-    )
-
-    response = client.put(
-        f"/api/v1.0/users/{user.id!s}/",
-        new_user_values,
-        format="json",
-    )
-
-    assert response.status_code == HTTP_200_OK
-    user.refresh_from_db()
-    user_values = dict(serializers.UserSerializer(instance=user).data)
-    for key, value in user_values.items():
-        if key in ["language", "timezone"]:
-            assert value == new_user_values[key]
-        else:
-            assert value == old_user_values[key]
-
-
-def test_api_users_update_authenticated_other():
-    """Authenticated users should not be allowed to update other users."""
-    user = factories.UserFactory()
-
-    client = APIClient()
-    client.force_login(user)
-
-    user = factories.UserFactory()
-    old_user_values = dict(serializers.UserSerializer(instance=user).data)
-    new_user_values = serializers.UserSerializer(instance=factories.UserFactory()).data
-
-    response = client.put(
-        f"/api/v1.0/users/{user.id!s}/",
-        new_user_values,
-        format="json",
-    )
-
-    assert response.status_code == HTTP_403_FORBIDDEN
-    user.refresh_from_db()
-    user_values = dict(serializers.UserSerializer(instance=user).data)
-    for key, value in user_values.items():
-        assert value == old_user_values[key]
-
-
-def test_api_users_patch_anonymous():
-    """Anonymous users should not be able to patch users via the API."""
-    user = factories.UserFactory()
-
-    old_user_values = dict(serializers.UserSerializer(instance=user).data)
-    new_user_values = dict(
-        serializers.UserSerializer(instance=factories.UserFactory()).data
-    )
-
-    for key, new_value in new_user_values.items():
-        response = APIClient().patch(
-            f"/api/v1.0/users/{user.id!s}/",
-            {key: new_value},
-            format="json",
-        )
-        assert response.status_code == HTTP_401_UNAUTHORIZED
-        assert response.json() == {
-            "detail": "Authentication credentials were not provided."
-        }
-
-    user.refresh_from_db()
-    user_values = dict(serializers.UserSerializer(instance=user).data)
-    for key, value in user_values.items():
-        assert value == old_user_values[key]
-
-
-def test_api_users_patch_authenticated_self():
-    """
-    Authenticated users should be able to patch their own user but only "language"
-    and "timezone" fields.
-    """
-    user = factories.UserFactory()
-
-    client = APIClient()
-    client.force_login(user)
-
-    old_user_values = dict(serializers.UserSerializer(instance=user).data)
-    new_user_values = dict(
-        serializers.UserSerializer(instance=factories.UserFactory()).data
-    )
-
-    for key, new_value in new_user_values.items():
-        response = client.patch(
-            f"/api/v1.0/users/{user.id!s}/",
-            {key: new_value},
-            format="json",
-        )
-        assert response.status_code == HTTP_200_OK
-
-    user.refresh_from_db()
-    user_values = dict(serializers.UserSerializer(instance=user).data)
-    for key, value in user_values.items():
-        if key in ["language", "timezone"]:
-            assert value == new_user_values[key]
-        else:
-            assert value == old_user_values[key]
-
-
-def test_api_users_patch_authenticated_other():
-    """Authenticated users should not be allowed to patch other users."""
-    user = factories.UserFactory()
-
-    client = APIClient()
-    client.force_login(user)
-
-    other_user = factories.UserFactory()
-    old_user_values = dict(serializers.UserSerializer(instance=other_user).data)
-    new_user_values = dict(
-        serializers.UserSerializer(instance=factories.UserFactory()).data
-    )
-
-    for key, new_value in new_user_values.items():
-        response = client.put(
-            f"/api/v1.0/users/{other_user.id!s}/",
-            {key: new_value},
-            format="json",
-        )
-        assert response.status_code == HTTP_403_FORBIDDEN
-
-    other_user.refresh_from_db()
-    user_values = dict(serializers.UserSerializer(instance=other_user).data)
-    for key, value in user_values.items():
-        assert value == old_user_values[key]
-
-
-def test_api_users_delete_list_anonymous():
-    """Anonymous users should not be allowed to delete a list of users."""
-    factories.UserFactory.create_batch(2)
-
-    client = APIClient()
-    response = client.delete("/api/v1.0/users/")
-
-    assert response.status_code == HTTP_401_UNAUTHORIZED
-    assert models.User.objects.count() == 2
-
-
-def test_api_users_delete_list_authenticated():
-    """Authenticated users should not be allowed to delete a list of users."""
-    user = factories.UserFactory()
-    factories.UserFactory.create_batch(2)
-
-    client = APIClient()
-    client.force_login(user)
-
-    response = client.delete(
-        "/api/v1.0/users/",
-    )
-
-    assert response.status_code == HTTP_405_METHOD_NOT_ALLOWED
-    assert models.User.objects.count() == 3
-
-
-def test_api_users_delete_anonymous():
-    """Anonymous users should not be allowed to delete a user."""
-    user = factories.UserFactory()
-
-    response = APIClient().delete(f"/api/v1.0/users/{user.id!s}/")
-
-    assert response.status_code == HTTP_401_UNAUTHORIZED
-    assert models.User.objects.count() == 1
-
-
-def test_api_users_delete_authenticated():
-    """
-    Authenticated users should not be allowed to delete a user other than themselves.
-    """
-    user, other_user = factories.UserFactory.create_batch(2)
-
-    client = APIClient()
-    client.force_login(user)
-
-    response = client.delete(f"/api/v1.0/users/{other_user.id!s}/")
-
-    assert response.status_code == HTTP_405_METHOD_NOT_ALLOWED
-    assert models.User.objects.count() == 2
-
-
-def test_api_users_delete_self():
-    """Authenticated users should not be able to delete their own user."""
-    user = factories.UserFactory()
-
-    client = APIClient()
-    client.force_login(user)
-
-    response = client.delete(
-        f"/api/v1.0/users/{user.id!s}/",
-    )
-
-    assert response.status_code == HTTP_405_METHOD_NOT_ALLOWED
-    assert models.User.objects.count() == 1
